backend/requirements.txt
google-cloud-storage
Flask
alembic
click
Flask-Cors
flask-marshmallow
Flask-Migrate
Flask-SQLAlchemy
greenlet
gunicorn
itsdangerous
Jinja2
Mako
MarkupSafe
marshmallow
marshmallow-sqlalchemy

python-dateutil
python-editor
six
SQLAlchemy
Werkzeug